requirements.txt
fastapi
uvicorn
jinja2
python-multipart
transformers
torch
sentence-transformers
langchain>=0.2.2
langchain-community>=0.0.1
langchain-huggingface
langchain-chroma
accelerate>=0.26.0
bitsandbytes
python-dotenv
chromadb
python-magic-bin
unstructured[pdf]
sentence-transformers
langchain-community[ollama]
langchain_ollama
